requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.28.78
botocore==1.31.78
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.14.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.3
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.9
